requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-pinecone
pinecone-client
python-dotenv
pypdf
